practice create task
made some finishing touches. Edited the appear on food code because i made the same mistake as the is colliding funtion by using two for loops. Also edited the check edges for the food clas to see if i could keep the food to appear on the screen if it loaded off screen